pyproject.toml
[project]
name = "salesflow-data-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "asn1crypto==1.5.1",
    "azure-common==1.1.28",
    "azure-core==1.33.0",
    "azure-identity==1.21.0",
    "azure-mgmt-core==1.5.0",
    "azure-mgmt-datafactory==9.1.0",
    "azure-mgmt-resource==23.3.0",
    "azure-mgmt-storage==22.1.1",
    "azure-storage-blob==12.25.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "cryptography==44.0.2",
    "filelock==3.18.0",
    "idna==3.10",
    "isodate==0.7.2",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "msal==1.32.0",
    "msal-extensions==1.3.1",
    "packaging==25.0",
    "platformdirs==4.3.7",
    "pycparser==2.22",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pyopenssl==25.0.0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "requests==2.32.3",
    "rich==14.0.0",
    "six==1.17.0",
    "snowflake-connector-python==3.14.0",
    "sortedcontainers==2.4.0",
    "tomlkit==0.13.2",
    "typing-extensions==4.13.2",
    "urllib3==2.4.0",
]

[dependency-groups]
dev = ["black>=25.1.0", "pytest>=8.3.5", "ruff>=0.11.5"]

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4

[tool.ruff.lint]
ignore = [
    "D",       # pydocstyle docstring rules
    "B905",    # `zip()` without an explicit `strict=` parameter
    "PLR0913", # Too many arguments to function call
    "TRY003",  # Avoid specifying long messages outside the exception class
    "EM101",   # Exception must not use a string literal, assign to variable first
    "EM102",   # Exception must not use an f-string literal, assign to variable first
    "BLE001",  # Do not catch blind exception: `Exception`
]
select = ["ALL"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"] # Allow assertions in tests

[project.scripts]
salesflow = "salesflow_data_pipeline.main:main"
salesflow-upload = "salesflow_data_pipeline.scripts.azure_blob_upload:main"
salesflow-init-db = "salesflow_data_pipeline.scripts.init_snowflake_db:main"
salesflow-create-pipeline = "salesflow_data_pipeline.scripts.adf_pipeline_creator:main"
